Disabled the python's GC in the server and added some cleanup.
Disabled the garbage collector (GC) of python in the server. This is a
risky movement but also an optimization.
Because there aren't any cyclic reference this should be ok. Eventually
the references to the objects will be drop and when the count of
reference reachs to zero the object will be destroyed.

Also filtered the list of endpoints before sending the message avoiding
sending a message to a disconnected endpoint.