Saniizer func to escape unicode/strings. Non-strings are unchanged.
This function sanitize a list of elements where each element can be an
unicode sequence, a string or anything else.
For sthe first two cases, the function will escape any weird character
making the element suitable for loggin it.
Non-srings are left unchaged.

Modified the log lines in the publish_subscribe system to use this
function.