Copy the callbacks and avoid a deadlock
Instead of protecting the for-loop that calls to the callbacks, copy
them.
Then, outside of the lock, call the callbacks. This reduce the critical
section and also disable any possible deadlock between that lock and the
code executed by the callbacks.
So there is no risk in write nested publishs and subscribes.