Optimized the message with a JSON over JSON approach.
Instead of sending a single JSON object with the topic and the data
inside of it, the new protocol consists in sending a single JSON object
with the topic and another JSON inside of it that encapsules the data.
In the former, the data was an arbitrary object serialized as JSON and
was parsed every time by the notifier to retrieve the topic.
In the latter the double encoding means that the data is just a huge and
large string from the point of view if the first layer.
Parsing this is almost trivial, making the whole system faster; only the
final endpoints will parse the inner JSON to get the real data.