Config parser and several shortcuts
A custom config parser to load the default 'global.cfg' configuration
from a file and overwrite it with the user personal 'user.cfg'
configuration.
Created a Frozen config to offer a read-only configuration to others (in
theory this should be thread-safe).

Created some useful shortcuts to start / stop a daemon as well to send
requests and receive the responses to a GDB instance.