Added a destroy method to cleanup the collector
If the collector is used to wrap a pub-sub callback, it could happen
that if not all the events are consumed by the collector function (using
get_next), the collector may block. If this happen, the thread in
pub-sub that dispatches the events will block too and the pub-sub close
method, which does a join of the dispatcher thread, will block.
The 'destroy' method of the collector will force to discard any current
pending and any future event: eventually the collector will not block
anymore and the dispatcher and the whole universe will finish correctly.