Collector's get_next now has a timeout
More inefficient but much more practical, the get_next method of the
collector has now a timeout (5 secs by default). If no data is received
by this, an exception will be raised.
This was designed to be more useful in the tests, otherwise if for some
reason an event wasn't received, the get_next would block forever.