Used regexps to normalize bytes/unicode repr in doctest.
In python 2, str/bytes and unicodes are represented by 'foo' and u'bar'.
In python 3, bytes and str are like b'foo' and 'bar'.
This breaks all the doctests.
Used a regular expression to replace u'foo' and b'bar' to 'foo' and
'bar' in any python version. This is a hack and may not work always.