Added goodbye message to dismiss errors in a normal shutdown.
A normal shutdown consists in the endpoint sending a 'goodbye' message,
following of the shutdown of the tcp channel. Under this scenario, both
parties, the endpoint and the notifier will log an informational message
but no error will be log.
If the 'goodbye' message is missing, or, if it is sent but the tcp channel is
closed in the middle of a message the connection is considered closed
abnormally and an error will be log.
Also, it is an error to send anything after the 'goodbye' message.